+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## Added
+- Option to only consider the sources as resources of the data that is indexed.
+
+## Changed
+- Update dependencies.
+
 ## [2.0.1] - 2020-04-19
 
 ### Fixed